+++ gmw/admin.py
@@ -4,3 +4,4 @@ from . import models as gmw_models
 
 admin.site.register(gmw_models.GMW)
 admin.site.register(gmw_models.MonitoringTube)
+admin.site.register(gmw_models.Event)